refactored code and fixed Activites / POIs layers rendering behind Map bug
